requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==35.0.0
Django==3.1.13
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
filetype==1.0.7
idna==2.10
MarkupSafe==2.0.1
Pillow>=8.2.0
psycopg2-binary>=2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse>=0.4.2
typing-extensions==3.7.4.3
urllib3==1.26.5


